Slims ProfilePage state moving to new CreateCardSection component
