+++ client_web/src/containers/ProfilePage/index.js
@@ -2,16 +2,10 @@ import React from "react";
 import { connect } from "react-redux";
 import { card } from "actions/cardActions";
 import { profile } from "actions/profileActions";
-import {
-  getProfile,
-  getCardFormIsLoading,
-  selectUser,
-  selectCardFormVisibility,
-  selectFormPicks,
-} from "reducers/selectors";
+import { getProfile, selectUser } from "reducers/selectors";
 import ProfileHeader from "components/ProfileHeader";
 import CardList from "components/CardList";
-import CardForm from "components/CardForm";
+import CreateCardSection from "components/CreateCardSection";
 
 class ProfilePage extends React.Component {
   constructor(props) {
@@ -25,16 +19,6 @@ class ProfilePage extends React.Component {
     this.props.getAllCards(username);
   }
 
-  handleCreateCardSubmit = (localFormState) => {
-    const { formPicks, createCard } = this.props;
-    const picksStrippedIds = formPicks.map(({ _id, ...rest }) => rest);
-
-    createCard({
-      comments: localFormState.comments,
-      picks: picksStrippedIds,
-    });
-  };
-
   handleProfileUpdateSubmit = (values) =>
     this.props.updateProfile(values, this.handleSetProfileEditHidden);
 
@@ -46,15 +30,7 @@ class ProfilePage extends React.Component {
 
   render() {
     const { profileHeader, profileCards } = this.props.profile;
-
-    const {
-      cardFormVisibility,
-      showCreateForm,
-      hideCreateForm,
-      isLoading,
-      user,
-    } = this.props;
-
+    const { user } = this.props;
     const isOwnProfile = this.props.match?.params?.username === user.username;
 
     return (
@@ -67,37 +43,14 @@ class ProfilePage extends React.Component {
           handleSetProfileEditVisible={this.handleSetProfileEditVisible}
           handleSetProfileEditHidden={this.handleSetProfileEditHidden}
         />
-        <div className="max-w-6xl m-auto">
-          {isOwnProfile &&
-            (cardFormVisibility.createFromVisible ? (
-              <div className="rounded-lg p-4 m-2 border-2 border-blue-500 text-xs">
-                <div className="flex justify-between">
-                  <h5 className="font-bold">Create a picks post</h5>
-                  <button onClick={hideCreateForm}>Cancel</button>
-                </div>
-                <CardForm
-                  onSubmit={this.handleCreateCardSubmit}
-                  isLoading={isLoading}
-                />
-              </div>
-            ) : (
-              <button
-                aria-label="create-post"
-                onClick={showCreateForm}
-                className="m-auto my-2 rounded-lg bg-red-400 flex py-4 px-8 text-white text-xs"
-              >
-                + New Picks
-              </button>
-            ))}
-        </div>
-        {profileCards.cardStatus === "loading" ? (
-          <small>Loading cards...</small>
-        ) : (
-          <CardList
-            cards={profileCards.cards}
-            loggedInUsername={user?.username}
-          />
-        )}
+
+        {isOwnProfile && <CreateCardSection />}
+
+        <CardList
+          cards={profileCards.cards}
+          loggedInUsername={user?.username}
+          isLoading={profileCards.cardStatus === "loading"}
+        />
       </div>
     );
   }
@@ -105,18 +58,11 @@ class ProfilePage extends React.Component {
 
 const mapState = (state) => ({
   profile: getProfile(state),
-  isLoading: getCardFormIsLoading(state),
   user: selectUser(state),
-  cardFormVisibility: selectCardFormVisibility(state),
-  formPicks: selectFormPicks(state),
 });
 
 export default connect(mapState, {
   getProfileHeader: profile.getSummary.request,
   updateProfile: profile.updateSummary.request,
   getAllCards: card.getAll.request,
-  createCard: card.create.request,
-  showCreateForm: card.form.create.show,
-  hideCreateForm: card.form.create.hide,
-  clearEditable: card.form.edit.clear,
 })(ProfilePage);
